+++ optuna/visualization/parallel_coordinate.py
@@ -15,10 +15,10 @@ if type_checking.TYPE_CHECKING:
     from typing import Optional  # NOQA
 
     if is_available():
-        from optuna.visualization._imports import Figure  # NOQA
+        from optuna.visualization.plotly_imports import Figure  # NOQA
 
 if is_available():
-    from optuna.visualization._imports import go
+    from optuna.visualization.plotly_imports import go
 
 logger = get_logger(__name__)
